requirements.txt
streamlit
rembg
Pillow
numpy
streamlit-drawable-canvas